requirements.txt
faiss-cpu
fastapi
langchain
langchain_community
langchain-openai
langchain-upstage
langchain-chroma
langserve
sentence-transformers
sse-starlette
python-dotenv
predibase

openai
tiktoken
chromadb

pymupdf